feat(gams): ship the .gms corpus in the wheel + `discopt gams-verify`
So pip users (not just repo clones) can validate the link against their
own GAMS install:

- Move the .gms smoke corpus + manifest.json from python/tests/data/gams/
  into the package at python/discopt/gams/data/, and add them to
  tool.maturin.include so they ship in the wheel (verified: the built
  wheel contains discopt/gams/data/*.gms + manifest.json).
- New discopt.gams.verify module (data_dir / load_manifest / verify)
  holds the end-to-end logic; expose data_dir/load_manifest/verify.
- New `discopt gams-verify` CLI command (and `python -m discopt.gams.verify`).
- scripts/verify_gams_link.py is now a thin wrapper; `make gams-verify`
  calls the package module. Tests load the corpus via data_dir().
- docs: smoke-testing section updated (packaged corpus + pip-only flow).

No local build for users (prebuilt wheels); the corpus travels with the
package. ruff + mypy clean.
