+++ python/discopt/gams/data/lp_prodmix.gms
