+++ python/discopt/gams/data/manifest.json
