+++ python/discopt/gams/data/mip_knapsack.gms
