+++ python/discopt/gams/data/nlp_circle.gms
