+++ python/discopt/gams/data/nlp_rosenbrock.gms
